Refactored index.js to import modules to init commands, events, and client
